Round completion and calculation
Stores team result summaries and rankings for each round in the database
so it can be retrieved quickly without recalculating.

Uses Hangfire to run the calculation task in the background. This might
also be used for setting the switch from registration to running for the
season.

There is some guarding in the admin UI when a round is marked as
completed.

TaskLog entity added to log messages from background tasks.
A REST API to get TaskLogs is present, no UI yet.

Captains and Vice Captains are now stored in a separate table so that
they can be identified if any round needs to be recalculated.
Previously certain actions would cause the c/vc info to be incomplete
when trades were made.