Removed external login providers, improved payment error handling
The external providers worked ok, but enabling google, facebook, and
whatever else is a lot of work so I've removed them and just kept local
accounts.

Local accounts now let you register with and change your name, it's also
added to the personal data.

Registering a team now includes better error handling and a new page
when completed where information and a link to the receipt is available.

I've removed a bunch of stuff including migrations and a redundent db
context.